+++ R/doi.R
@@ -31,10 +31,11 @@ so_doi <- function(cache_directory, ...) {
 
 ## internal, default to this DOI if necessary
 ## this is the DOI of the latest version
-so_default_doi <- function() "10.26179/5d1aec22f41d5"
+so_default_doi <- function() "10.5281/zenodo.5072528"
 
 ## DOI versions
 ##
+## 5-Jul-2021 "10.5281/zenodo.5072528"
 ## 2-Jul-2019 "10.26179/5d1aec22f41d5"
 ## 21-Sep-2018 "10.26179/5ba3396f46e42"
 ## 9-Aug-2018 "10.26179/5b6cd40bb6935"