+++ tests/testthat/test-get-data.R
@@ -2,9 +2,10 @@ context("sohungry data retrieval")
 
 test_that("data retrieval works", {
     skip_on_cran()
-    expect_warning(xi <- so_isotopes(method="get"))
-    expect_warning(xi <- so_isotopes(method="get", format="wide"))
-    xi <- so_isotopes(method="get", format="mv")
+    xi <- so_isotopes(method="get")
+    expect_error(xi <- so_isotopes(method="get", format="wide"), regexp = "no longer supported")
+    xi2 <- so_isotopes(method="get", format="mv")
+    expect_identical(xi, xi2)
 
     xd <- so_diet(method="get")
 })